añadidos ejercicios AD y arreglado Chronometer de PUL/Buscaminas

+++ AD/apache-tomcat-7.0.23/webapps/ejemploMVC/WEB-INF/classes/ejemplo/Departamentos.java
@@ -1,17 +1,17 @@
 package ejemplo;
 public class Departamentos {
-	private int deptno;
+	private byte deptno;
 	private String dnombre;
 	private String loc;
     
 	public Departamentos() {}
-	public Departamentos(int deptno, final String dnombre, final String loc) {
+	public Departamentos(byte deptno, final String dnombre, final String loc) {
 		this.deptno = deptno;
 		this.dnombre = dnombre;
 		this.loc = loc;
 	}
-	public int getDeptno() {return this.deptno;}
-	public void setDeptno(int deptno) {this.deptno = deptno;}
+	public byte getDeptno() {return this.deptno;}
+	public void setDeptno(byte deptno) {this.deptno = deptno;}
 	public String getDnombre() {return this.dnombre;}
 	public void setDnombre(String dnombre) {this.dnombre = dnombre;}
 	public String getLoc() {return this.loc;}
